+++ @cybernaut/chrome/src/tests/Chrome.ts
@@ -33,6 +33,7 @@ function createUrl(name: string): string {
 }
 
 beforeEach(async () => {
+  /* The GUI mode can only be tested manually. */
   chrome = await Chrome.launch(true);
   port = await getPortPromise();
   server = app.listen(port);
@@ -53,7 +54,7 @@ describe('Chrome.scriptResult()', () => {
       .is.equalTo('ab');
 
     expect(condition.description).toBe(
-      "Chrome.launch(true).scriptResult((a, b) => a + b, 'a', 'b').is.equalTo('ab')"
+      `Chrome.launch(${chrome.headless}).scriptResult((a, b) => a + b, 'a', 'b').is.equalTo('ab')`
     );
   });
 
@@ -83,7 +84,7 @@ describe('Chrome.scriptResult()', () => {
     await expect(assert(condition)).rejects.toEqual(
       new Error(
         [
-          'Assert: Chrome.launch(true).scriptResult(() => {',
+          `Assert: Chrome.launch(${chrome.headless}).scriptResult(() => {`,
           "            throw new Error('an error');",
           '        }).is.equalTo(undefined) => Error: an error'
         ].join('\n')
@@ -101,7 +102,7 @@ describe('Chrome.scriptResult()', () => {
     await expect(assert(condition)).rejects.toEqual(
       new Error(
         [
-          'Assert: Chrome.launch(true).scriptResult(() => {',
+          `Assert: Chrome.launch(${chrome.headless}).scriptResult(() => {`,
           '            throw new Error();',
           '        }).is.equalTo(undefined) => Error: Unknown error'
         ].join('\n')
@@ -115,7 +116,7 @@ describe('Chrome.runScript()', () => {
     const action = chrome.runScript((a: string, b: string) => a + b, 'a', 'b');
 
     expect(action.description).toBe(
-      "Chrome.launch(true).runScript((a, b) => a + b, 'a', 'b')"
+      `Chrome.launch(${chrome.headless}).runScript((a, b) => a + b, 'a', 'b')`
     );
   });
 
@@ -143,7 +144,7 @@ describe('Chrome.runScript()', () => {
     await expect(perform(action)).rejects.toEqual(
       new Error(
         [
-          'Perform: Chrome.launch(true).runScript(() => {',
+          `Perform: Chrome.launch(${chrome.headless}).runScript(() => {`,
           "            throw new Error('an error');",
           '        }) => Error: an error'
         ].join('\n')
@@ -159,7 +160,7 @@ describe('Chrome.runScript()', () => {
     await expect(perform(action)).rejects.toEqual(
       new Error(
         [
-          'Perform: Chrome.launch(true).runScript(() => {',
+          `Perform: Chrome.launch(${chrome.headless}).runScript(() => {`,
           '            throw new Error();',
           '        }) => Error: Unknown error'
         ].join('\n')
@@ -173,7 +174,9 @@ describe('Chrome.navigateTo()', () => {
     const action = chrome.navigateTo(createUrl('navigateTo'), true);
 
     expect(action.description).toBe(
-      `Chrome.launch(true).navigateTo('${createUrl('navigateTo')}', true)`
+      `Chrome.launch(${chrome.headless}).navigateTo('${createUrl(
+        'navigateTo'
+      )}', true)`
     );
   });
 
@@ -191,7 +194,7 @@ describe('Chrome.pageTitle', () => {
     const condition = chrome.pageTitle.is.equalTo('pageTitle');
 
     expect(condition.description).toBe(
-      "Chrome.launch(true).pageTitle.is.equalTo('pageTitle')"
+      `Chrome.launch(${chrome.headless}).pageTitle.is.equalTo('pageTitle')`
     );
   });
 
@@ -207,7 +210,7 @@ describe('Chrome.pageUrl', () => {
     const condition = chrome.pageUrl.is.equalTo('pageUrl');
 
     expect(condition.description).toBe(
-      "Chrome.launch(true).pageUrl.is.equalTo('pageUrl')"
+      `Chrome.launch(${chrome.headless}).pageUrl.is.equalTo('pageUrl')`
     );
   });
 
@@ -223,7 +226,7 @@ describe('Chrome.captureScreenshot()', () => {
     const action = chrome.captureScreenshot(true);
 
     expect(action.description).toBe(
-      'Chrome.launch(true).captureScreenshot(true)'
+      `Chrome.launch(${chrome.headless}).captureScreenshot(true)`
     );
   });
 
@@ -270,7 +273,7 @@ describe('Chrome.emulateMobileDevice()', () => {
     const action = chrome.emulateMobileDevice(device, true);
 
     expect(action.description).toBe(
-      "Chrome.launch(true).emulateMobileDevice({width: 1920, height: 1080, pixelRatio: 3, userAgent: 'an user agent'}, true)"
+      `Chrome.launch(${chrome.headless}).emulateMobileDevice({width: 1920, height: 1080, pixelRatio: 3, userAgent: 'an user agent'}, true)`
     );
   });
 
@@ -324,7 +327,7 @@ describe('Chrome.emulateMobileDevice()', () => {
   });
 
   it('should scale the view to fit the available browser window area', async () => {
-    await perform(chrome.navigateTo(createUrl('emulateMobileDevice')));
+    await perform(chrome.navigateTo(createUrl('emulateMobileDevice'), true));
     await perform(chrome.emulateMobileDevice(device, true));
 
     const writeToFile = process.env.CI !== 'true';